requirements.txt
aiogram==3.7.0
python-dotenv
openpyxl
